Updating the Apple Silicon Branch
The docker image requires some small modifications for the scripts to work because the OSIRIS version that the docker image pulls generate data in a slightly different format.

Also assumes that the ARM 64 version will be run under Apple Silicon which has a minimum of 8 cores and start making the jobs parallel.